refactor(generator): ReaderMethod/WriterMethod.writeTo() for union methods (C32d Step 2)
Move union reader/writer method generation into method classes:
- ReaderMethod.writeTo() generates readUnionName(JsonNode, ModelType) with
  full variant dispatch (entity/primitive/entityList/primitiveList)
- WriterMethod.writeTo() generates writeUnionName(Union) with same dispatch

Stage reductions:
- CreateReadersStage: 578 → 415 (-163)
- CreateWritersStage: 513 → 404 (-109)

Added to CodeGenContext: getModelTypeEnumFQN(), resolveUnionPackage(),
made lookupJavaEntity() public.
